Splitted regions 1 to 3 and constants into separate modules
