+++ src/iapws97/constants.rs
@@ -1,0 +1,3 @@
+pub(crate) const _R: f64 = 0.461526;
+pub(crate) const T_C: f64 = 647.096;
+pub(crate) const RHO_C: f64 = 322.0;
